+++ pyslvs_ui/io/slvs/__init__.py
@@ -7,12 +7,12 @@ __copyright__ = "Copyright (C) 2016-2019"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
-from .read import SlvsParser2
+from .read import SlvsParser
 from .frame import slvs2_frame
 from .part import slvs2_part, boundary_loop
 
 __all__ = [
-    'SlvsParser2',
+    'SlvsParser',
     'slvs2_frame',
     'slvs2_part',
     'boundary_loop',