+++ pyslvs_ui/io/slvs/read.py
@@ -7,11 +7,11 @@ __copyright__ = "Copyright (C) 2016-2019"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
-from typing import Tuple, List, Dict
+from typing import Tuple, List, Set, Dict
 from pyslvs import VLink
 
 
-class SlvsParser2:
+class SlvsParser:
 
     """Use to read data from solvespace file format."""
 
@@ -78,8 +78,8 @@ class SlvsParser2:
                 # 0x4 << 16 == 0x40000
                 requests.append(int16(request['Request.h.v']) << 16)
 
-        vlinks = {link: {link + 1, link + 2} for link in requests}
-        vlinks[0] = set()
+        links: Dict[int, Set[int]] = {link: {link + 1, link + 2} for link in requests}
+        links[0] = set()
 
         # Constraint: Adjacency
         # Grounded. Other links at least will greater than 4
@@ -88,42 +88,41 @@ class SlvsParser2:
                 continue
             if constraint['Constraint.type'] == '20':
                 now_replace = int16(constraint['Constraint.ptA.v'])
-                for vlink in vlinks.values():
+                for vlink in links.values():
                     num = int16(constraint['Constraint.ptB.v'])
                     if num in vlink:
                         vlink.remove(num)
                         vlink.add(now_replace)
             elif constraint['Constraint.type'] == '31':
-                vlinks[0].add(int16(constraint['Constraint.ptA.v']))
+                links[0].add(int16(constraint['Constraint.ptA.v']))
 
         pos = {}
-        for vlink in vlinks.values():
+        for vlink in links.values():
             for point in vlink:
                 if point not in pos:
-                    pos[point] = None
-        points = sorted(pos)
+                    pos[point] = (0., 0.)
 
         # Entities: Get positions
         for entity in self.entities:
-            if entity['Entity.type'] == '2001':
-                num = int16(entity['Entity.h.v'])
-                if num in pos:
-                    pos[num] = (
-                        entity['Entity.actPoint.x'],
-                        entity['Entity.actPoint.y'],
-                    )
+            if entity['Entity.type'] != '2001':
+                continue
+            num = int16(entity['Entity.h.v'])
+            if num in pos:
+                pos[num] = (
+                    entity.get('Entity.actPoint.x', 0.),
+                    entity.get('Entity.actPoint.y', 0.),
+                )
 
         # Rename link names
-        for i, name in enumerate(sorted(vlinks)):
+        vlinks = {}
+        for i, name in enumerate(sorted(links)):
             if i == 0:
-                vlinks[VLink.FRAME] = vlinks.pop(name)
+                vlinks[VLink.FRAME] = links.pop(name)
             else:
-                vlinks[f'link_{i - 1}'] = vlinks.pop(name)
-
+                vlinks[f'link_{i - 1}'] = links.pop(name)
         exprs = []
-        for num in points:
+        for num in sorted(pos):
             x, y = pos[num]
-            links = [name for name, link in vlinks.items() if (num in link)]
-            exprs.append(f"J[R, color[Green], P[{x}, {y}], L[{', '.join(links)}]]")
-
+            links_str = ", ".join(name for name, link in vlinks.items() if num in link)
+            exprs.append(f"J[R, color[Green], P[{x}, {y}], L[{links_str}]]")
         return "M[" + ", ".join(exprs) + "]"