+++ pyslvs_ui/main_window/io.py
@@ -46,7 +46,7 @@ from pyslvs_ui.info import (
 from pyslvs_ui.io import (
     ScriptDialog,
     slvs_process_script,
-    SlvsParser2,
+    SlvsParser,
     SlvsOutputDialog,
     DxfOutputDialog,
     OverviewDialog,
@@ -77,13 +77,13 @@ class IOMethodInterface(ActionMethodInterface, ABC):
                         yield (vlink.points[i - 1], p)
         return func
 
-    def __read_slvs2(self, file_name: str) -> None:
+    def __read_slvs(self, file_name: str) -> None:
         """Read slvs format.
 
         + Choose a group.
         + Read the entities of the group.
         """
-        parser = SlvsParser2(file_name)
+        parser = SlvsParser(file_name)
         if not parser.is_valid():
             QMessageBox.warning(self, "Format error", "The format is not support.")
             return
@@ -312,7 +312,7 @@ class IOMethodInterface(ActionMethodInterface, ABC):
         if not file_name:
             file_name = self.input_from("project", [
                 "Pyslvs project (*.pyslvs)",
-                "Solvespace 2.x (*.slvs)",
+                "Solvespace 2.x-3.x (*.slvs)",
             ])
             if not file_name:
                 return
@@ -321,7 +321,7 @@ class IOMethodInterface(ActionMethodInterface, ABC):
         if suffix == 'pyslvs':
             self.project_widget.read(file_name)
         elif suffix == 'slvs':
-            self.__read_slvs2(file_name)
+            self.__read_slvs(file_name)
         else:
             QMessageBox.warning(
                 self,
@@ -640,7 +640,7 @@ class IOMethodInterface(ActionMethodInterface, ABC):
         if suffix == 'pyslvs':
             self.project_widget.read(ARGUMENTS.filepath)
         elif suffix == 'slvs':
-            self.__read_slvs2(ARGUMENTS.filepath)
+            self.__read_slvs(ARGUMENTS.filepath)
         else:
             QMessageBox.warning(
                 self,